Added fronting and backing feature.
This commit contains changes to the GUI. It is now possible to choose
what width units to use for the different components. An option to not
include absorber has been included. Also the option to have backing and
fronting on the target has been included. These updates has not been
implemented in the backend where the actual calculations are done.