Added the possibility of running batch files.
Added the possibility of running batch files. The program is now
capable of performing calculations on inverse kinematics using batch
files. The program can also use custom stopping powers provided as a
tabulated file. The output of the batch is the excitation energy as a
function of the energy deposited in the particle telescope. The batch
file is a text file and needs to get a file containing the angles for
which is should calculate the scattering (in radians). The file can
have one or two indexes before each angle that will be outputted in the
result file. Also added some user feedback for the progress of the
calculations.